refactor(groq): Allow defineQuery to live in the groq package
