+++ packages/groq/package.json
@@ -26,23 +26,16 @@
   "license": "MIT",
   "author": "Sanity.io <hello@sanity.io>",
   "sideEffects": false,
+  "type": "module",
   "exports": {
     ".": {
-      "source": "./src/groq.ts",
-      "import": "./lib/groq.mjs",
-      "require": "./lib/groq.js",
+      "require": "./lib/groq.cjs",
       "default": "./lib/groq.js"
-    },
-    "./define": {
-      "source": "./src/define.ts",
-      "import": "./lib/define.mjs",
-      "require": "./lib/define.js",
-      "default": "./lib/define.js"
     },
     "./package.json": "./package.json"
   },
-  "main": "./lib/groq.js",
-  "module": "./lib/groq.esm.js",
+  "main": "./lib/groq.cjs",
+  "module": "./lib/groq.js",
   "types": "./lib/groq.d.ts",
   "files": [
     "lib",
