+++ packages/groq/src/_exports.mts.ts
@@ -1,0 +1,2 @@
+export {defineQuery} from './define'
+export {groq as default} from './groq'
